Moved position tests to own package
